fix: react-query 받아오는 코드 수정 , 추후 mypage 받아오는 부분 undefined 해결 필요
