+++ src/hooks/queries/user/useGetMypage.tsx
@@ -1,28 +1,26 @@
 import { useQuery } from '@tanstack/react-query';
-import { instance } from '../../../apis/client';
+import { get } from '../../../apis/client';
 import { MyPageType } from '../../../types/user';
-import { useRecoilValue } from 'recoil';
-import { memberIdState } from '../../../recoil/atoms';
 
 interface MyPageResponse extends Response {
   data: MyPageType;
 }
 
-export const MYPAGE_BY_ID_QUERY_KEY: string[] = ['myPageById'];
+export const MYPAGE_BY_ID_QUERY_KEY: string[] = ['myPageData'];
 
-export const fetchMyPage = (): Promise<MyPageResponse> =>
-  instance.get(`/member/mypage/2`).then((res) => res.data);
+export const fetchMyPage = async (): Promise<MyPageResponse> => get(`/member/mypage`);
 
 const useGetMyPage = () => {
-  const userId: number = useRecoilValue(memberIdState);
+  const shouldFetch = !!localStorage.getItem('EXIT_LOGIN_TOKEN');
 
-  if (localStorage.getItem('EXIT_LOGIN_TOKEN')) {
-    const { data } = useQuery({
-      queryKey: MYPAGE_BY_ID_QUERY_KEY,
-      queryFn: () => fetchMyPage().then((res) => res.data),
-    });
-    return data;
-  }
+  const { data } = useQuery({
+    queryKey: MYPAGE_BY_ID_QUERY_KEY,
+    queryFn: () => fetchMyPage(),
+    enabled: shouldFetch,
+  });
+
+  console.log(data);
+  return data?.data;
 };
 
 export default useGetMyPage;
